An S4 Class to Represent a Trace.

Defines the trace class, which contains all characteristics of a trace, including the setting, the agents at each iteration, and the time between each iteration.

Slots

id

Character that serves as an identifier for the trace.

time_step

Numerical denoting the time that passes at each iteration in seconds.

setting

Object of the background-class.

states

List of lists containing the state of the agent-classs at each iteration.

variables

List of list of user-specified variables that are used to control the simulation (see the fx argument of simulate) at each iteration.

See also