An S4 Class to Represent Agents.

Defines the `agent` class, which contains all characteristics of an agent. Some of these characteristics are time-independent (e.g., parameters and set of goals) while others change during each time step in the simulation or estimation (e.g., speed and orientation).

Slots

id

A numerical index for the agent

position

A numerical vector of two elements denoting the position of the agent in 2D space

size

A numeric indicating the size of the radius of the agent

orientation

A numeric denoting the orientation of the agent in degrees

interacting

A logical denoting whether the agent is interacting with something in its environment

parameters

A named list that contains the individual-specific parameters for the agent

goals

A list containing the id's of the goals that are assigned to the agent

status

A character that denotes the status of the agent, or what the agent is doing at the moment. Can take on `"move"`, `"stop"`, `"reorient"`, and `"exit"`

See also