Transform segments to polygons

This function is used to transform an object of the segment-class to an object of the polygon-class. Done as a precomputation in the constructor for the background-class (filling the precomputed_limited_access slot based on the limited_access slot.

Usage

compute_limited_access(segment)

Arguments

segment

Object of the segment-class.

Value

Object of the polygon-class.

See also