Compute utility variables

Rcpp version of the compute_utility_variables function.

compute_utility_variables_rcpp(
  agent,
  state,
  background,
  agent_specifications,
  centers,
  check_original
)

Arguments

agent

Object of the agent-class.

state

Object of the state-class.

background

Object of the background-class.

agent_specifications

List created by the create_agent_specifications function. Contains all information of all agents within the current state and allows for the communication between the predped simulation functions and the m4ma utility functions.

centers

Numerical matrix containing the coordinates at each position the object can be moved to. Should have one row for each cell.

check_original

Logical matrix of dimensions 11 x 3 denoting whether an agent can move to a given cell (TRUE) or not (FALSE).

Value

Data.frame containing all of the needed variables to be able to compute the values of the utility functions.

See also