Create agent specifications

This list translates the information available in the agents slot of the current status of the state-class to a list with all this information in numeric vectors or matrices instead of inside objects. Allows for a translation from the object-oriented way of doing things in predped to the vectorized way of doing things in m4ma.

Usage

create_agent_specifications(
  agent_list,
  stay_stopped = TRUE,
  time_step = 0.5,
  cpp = TRUE
)

Arguments

agent_list

List of objects of the agent-class.

stay_stopped

Logical denoting whether agents will predict others that are currently not moving to remain immobile in the next iteration. Defaults to TRUE.

time_step

Numeric denoting the number of seconds each discrete step in time should mimic. Defaults to 0.5, or half a second.

cpp

Logical denoting whether to use the Rcpp alternative (TRUE) or the R alternative of this function (FALSE). Defaults to TRUE.

Value

List containing all information of all agents within the current state.

See also