An S4 Class to Represent the M4MA model.

The `predped` class defines the M4MA model that can be used to simulate data from. It contains the setting in which the agents should walk around as well as the parameters of those agents. At this moment, only supports the use of archetypes: Prototypical agents that have a given parameter set.

Slots

id

A character that can be used to identify the specific model defined by the class

setting

A dataframe containing the objects that define the setting in which the agents will walk around

parameters

A dataframe containing the parameters per archetype in its columns. Defaults to the `params_archetype` dataframe.

archetypes

A vector of names for the archetypes one wants to use in the simulation. Defaults to all archetypes in the `params_archetype` dataframe.

weights

A numeric vector containing the probability with which an agent can be of a given archetype. The weights should be in the same order as the `archetypes` argument. Defaults to an equal weighting of all archetypes in `params_archetype`