Transform to positive axis

Use the real (normal, unbounded) definition of the parameters and transform them to a positive (probit, bounded) scale.

Usage

to_bounded(parameters, bounds)

Arguments

parameters

Dataframe or named list containing values for the parameters of M4MA.

bounds

Named numeric matrix containing the bounds for the parameters of M4MA.

Value

Dataframe or named list containing the transformed parameters

See also