Transform data to a trace

This function does the opposite of unpack_trace. It takes in a data.frame and return a trace according to predpeds requirements. The data.frame should at least have the column names "x", "y", "time", and "id", containing the coordinates, times at which the data were gathered (in seconds), and the id-number of the person whose data it is. Additionally, data.frame needs information on the goals that agents were trying to achieve at each timepoint, of which their positions should be saved under "goal_x" and "goal_y", and their id to "goal_id".

to_trace(
  data,
  background,
  b_turning = NULL,
  a_turning = NULL,
  velocities = c(1.5, 1, 0.5),
  orientations = c(72.5, 50, 32.5, 20, 10, 0, -10, -20, -32.5, -50, -72.5),
  time_step = 0.5,
  threshold = qnorm(0.975, 2 * 0.035, 4 * 0.035^4)/time_step,
  stay_stopped = TRUE,
  cpp = TRUE,
  ...
)

Arguments

data

Instance of a data.frame containing the data you want to transform.

background

Instance of the background-class containing the setting in which the data were gathered.

...

Arguments passed to add_cells.

Examples

# This is my example