Transform to real axis

Use the probit (positive, bounded) definition of the parameters and transform them to a real (normal, unbounded) scale.

to_unbounded(parameters, bounds)

Arguments

parameters

Dataframe or named list containing values for the parameters of M4MA.

bounds

Named numeric matrix containing the bounds for the parameters of M4MA.

Value

Dataframe or named list containing the transformed parameters

See also