Move an Agent

Move an Agent

update_position(
  agent,
  state,
  agent_specifications,
  background,
  nests = list(Central = c(0, 6, 17, 28), NonCentral = c(0:33)[-c(6, 17, 28)], acc =
    c(1:11), const = c(12:22), dec = c(0, 23:33)),
  alpha = list(Central = rep(1/3, 4), NonCentral = c(1/3, rep(0.5, 4), 1/3, rep(0.5, 9),
    1/3, rep(0.5, 9), 1/3, rep(0.5, 5)), acc = c(rep(0.5, 4), 1, 1/3, rep(0.5, 5)), const
    = c(rep(0.5, 4), 1, 1/3, rep(0.5, 5)), dec = c(1/3, rep(0.5, 4), 1, 1/3, rep(0.5,
    5))),
  velocities = matrix(rep(c(1.5, 1, 0.5), each = 11), ncol = 3),
  orientations = matrix(rep(c(72.5, 50, 32.5, 20, 10, 0, 350, 340, 327.5, 310, 287.5),
    times = 3), ncol = 3),
  standing_start = 0.1,
  time_step = 0.5,
  report = TRUE
)

Arguments

agent

The agent to move

state

The current state of affairs. Importantly, the agent in `agent` should not be present in that state

background

The setting in which agents are walking around

velocities

Matrix that contains the change in velocity per cell that the agent might move to. Defaults to an 11 by 3 matrix where each row contains 0.5 (deceleration), 1 (same speed), and 1.5 (acceleration).

orientations

Matrix that contains the change in orientation per cell that the agent might move to. Defaults to an 11 by 3 matrix where each column contains 72.5, 50, 32.5, 20, 10, 0, 350, 340, 327.5, 310, and 287.5 degrees.

standing_start

Numeric denoting the speed of the agent when they resume walking after stopping. Defaults to `0.1`

time_step

Numeric denoting the time step taken between iterations in seconds. Defaults to `0.5` or half a second.

agent_predictions

A list containing the predictions of where each agent might move to

Value

Updated agent

See also